Add router to route requests to handlers
The router iterates through a list of all handlers for a matching path
between the handler and given request. If a match is found the handler
generates the response for the request. If no match is found the default
handler (404 Not Found) is engaged to send the response.

In upcoming features an "allowed methods" field will be added to each
handler stating what methods they can respond to.