Create handlers to individualize responses
A handler is created for each path that the server is designated to
respond to, and will build the correct response when passed a request
from the router.

In subsequent iterations handlers will also be able to know which HTTP
methods they should respond to, otherwise they'll send a 405 method not
allowed response.