Send response status line to client
The response object is currently comprised of just a StatusLine object.
In subsequent features which implement headers and body functionality the
Response object will be comprised of a StatusLine object, a Headers
object, and a Body object, which will either be stringified or converted
to a byte stream before being sent to the client.