Define CL args and validation
Args defines --port as a required argument when starting the program. It
is set to accept an integer between 1024 and 49151 inclusive so as not
to interfere with system ports (under 1024) and dynamic ports (over
49151). If no argument is provided it will display usage to the user.
If a non integer is given it will exit and display a prompt to the user.
Users can also pass the --help argument to get the usage.