Convert Handler to functional interface
Handler is a functional interface in order to take advantage of lambdas
as callbacks. A Handler is created to give custom functionality to
responding to routes defined by the application.

When a handler is invoked it sends the request (and body if applicable"
to HandlerHelper, which creates a new ResponseBuilder and sends the
necessary information to build the correct response.