Separate application from server
App contains defined routes, which are built in the Routes class. The
routes are stored as a TreeMap in order to easily accommodate an unknown
number of routes. The routes in the TreeMap correlate with a HashMap
that contains the allowed methods and the corresponding handler.

Once the routes are defined, App is injected into the Server to allow
for the server to check incoming requests against known routes housed in
the App.