Add PathFinder to create response for request path
PathFinder checks the incoming request for a matching path in the routes
defined by the app. If there's a match, the response creation will be
delegated to the handler that matches the found path/route.

If a match is not found then the response creation gets handed to
ResponseBuilder to return a 404 Not Found Response. Similarly, if a
matching path is found but is being accessed by an incorrect method,
ResponseBuilder will return a Method Not Allowed response.