Add acceptance tests to Travis CI
Acceptance tests are run by a script in .travis.yml that will launch an
executable shell file that will get the acceptance test suite from
GitHub and run the test command. This was done because the acceptance
test repository is private, so I cloned it and relocated the remote to a
repository on my GitHub account.