Chunk input stream to get full request
Formerly input was received by calling readline() on the BufferedReader
that read from the clientSocket. This was working fine until the program
needed to access the body of the request.

Now readline() has been replaced with chunkStream(), which takes a bite
out of the incoming request stream with a character buffer. If the
incoming stream is empty it will throw an exception.