Fix redirect functionality
There's a weird thing that happens with the acceptance test suite, where
for some reason the request header does not contain the full ip address
+ port, it just contains the ip address. For clients other than the
acceptance suite, the redirect Location header appends the redirect path
to the host of the request. If the request comes from the acceptance
suite, :5000 is appended to the host. It's a mystery to me as to why the
acceptance suite request headers don't contain the full Host, yet
require the full host in order for the test to pass.