Address redirect bug
The issue occurred when trying to parse a request that came originated
from localhost. The request parser was originally splitting the headers
section by searching for a colon (:) as the middle of a header key/value
pair. Thus when a request came in from localhost, the host field would
have two colons (Host: localhost:5000) and the port number would be
discarded.

The fix is for the header parsing to split the headers section by a
colon plus a single space (: ) instead of just a colon. That way if the
host is localhost it will correctly parse the port number.