Reject empty requests
The first change was done in Client, where chunkStream() will now loop
until the bufferedReader is not ready (meaning there's no more data to
be read). Since this always returns a string (an empty string if the
request was empty) the run() method in ServerRunner now checks to see if
the incoming request is empty or not, where previous it was checking for
a null value.